Renamed the top-level package to `term_image`
- Change: Unified the top-level package name with the project name.
