+++ term_image/exceptions.py
